chore: adds markdown rendering
