+++ sample/src/pages/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import g from 'glamorous'
 import Link from 'gatsby-link'
+import Specs from '../components/Specs'
 import { rhythm } from '../utils/typography'
 
 const toPost = edge => ({
@@ -21,27 +21,6 @@ const Post = ({ slug, title, date, excerpt }) => (
     </Link>
   </div>
 )
-
-const Spec = ({ name, title }) => (
-  <li>
-    <Link to={`/apis/${name}`}>{title}</Link>
-  </li>
-)
-
-Spec.propTypes = {
-  name: PropTypes.string.isRequired,
-  title: PropTypes.string,
-}
-
-const Specs = ({ specs }) => (
-  <div>
-    <ul>{specs.map(s => <Spec key={s.id} name={s.name} title={s.title} />)}</ul>
-  </div>
-)
-
-Specs.propTypes = {
-  specs: PropTypes.array.isRequired,
-}
 
 export default ({ data }) => {
   const postCount = data.allMarkdownRemark.totalCount
